requirements.txt
aiohttp==3.9.1
aiosmtplib==3.0.1
beautifulsoup4==4.12.2
pandas==2.1.4
requests==2.31.0
lxml==4.9.4
email-validator==2.1.0
python-dotenv==1.0.0
